feat: add --no-version-check flag
Currently whenever a local tool version doesn't match the version
in a baseline file the developers are forced to update it
even when there are no secrets change. Together with an updated timestamp
this creates a constant merge conflict problems when multiple developers
are working in parallel.

This change adds a new flag `--no-version-check` that allows to skip
the version update in the baseline file when there are no other changes.

The new flag is false by default, allowing to keep back compatibility
with the previous behaviour.

Signed-off-by: Eric Avdey <eiri@eiri.ca>